backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
environs==14.1.1
fastapi==0.115.0
fastapi_cors==0.0.6
h11==0.14.0
httpcore==1.0.8
httpx==0.27.2
idna==3.10
marshmallow==4.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.13.2
uvicorn==0.30.6
